alerts/requirements.txt
confluent-kafka==2.3.0
redis==5.0.1
requests==2.31.0
psycopg2-binary==2.9.9
prometheus-client==0.19.0
python-dotenv==1.0.0
